stick pretty formatter in its own module
